requirements.txt
aiohttp==3.7.4.post0
aniso8601==7.0.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
cftime==1.4.1
channels==2.4.0
chardet==4.0.0
charset-normalizer==2.1.1
cloudinary==1.25.0
constantly==15.1.0
coverage==5.1
cryptography==3.4.7
cssselect==1.1.0
daphne==2.5.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.15
django-channels-graphql-ws==0.5.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-graphql-geojson==0.1.4
django-graphql-jwt==0.3.1
django-graphql-social-auth==0.1.4
djangorestframework==3.11.0
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.1.1
flake8==3.8.2
graphene==2.1.8
graphene-django==2.10.1
graphene-django-extras==0.4.9
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==21.3.0
lml==0.1.0
lxml==4.6.3
markdown2==2.4.0
mccabe==0.6.1
more-itertools==8.7.0
msgpack==0.6.2
multidict==5.1.0
mypy==0.981
mypy-extensions==0.4.3
nanoid==2.0.0
numpy==1.21.6
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyexcel==0.6.6
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.4
pyexcel-ods3==0.6.0
pyexcel-xlsx==0.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyquery==1.4.3
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2020.1
readtime==1.1.1
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2.1
sqlparse==0.3.1
text-unidecode==1.3
texttable==1.6.3
tomli==2.0.1
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.1.0
yarl==1.6.3
zipp==3.1.0
zope.interface==5.4.0
